requirements.txt
flask>=3.0
openpyxl>=3.1
docxtpl>=0.17
python-docx>=1.1
werkzeug>=3.0
